// src/generated/model/studentDetailResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface StudentDetailResponse {
  id?: string;
  studentNo?: string;
  fullName?: string;
  dateOfBirth?: string;
  gender?: string;
  nationality?: string;
  placeOfBirth?: string;
  religion?: string;
  passportNumber?: string;
  passportExpiryDate?: string;
  telephoneNumber?: string;
  email?: string;
  currentResidentialAddress?: string;
  permanentResidentialAddress?: string;
  educationBackground?: string;
  fatherName?: string;
  motherName?: string;
  emergencyContactName?: string;
  emergencyContactRelationship?: string;
  emergencyContactPhone?: string;
  status?: string;
  remarks?: string;
  createdAt?: string;
  updatedAt?: string;
}